mechanisms/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mechanisms"
authors = [{name = "Robin De Schepper", email = "robingilbert.deschepper@unipv.it"}]
readme = "README.rst"
license = {file = "LICENSE"}
dynamic = ["version", "description"]
dependencies = ["nmodl-glia>=4.0.0b0,<=4.0.0b9999"]

[project.entry-points."glia.package"]
"mechanisms" = "mechanisms:package"

[tool.glia.package]
mod-dir = "mods"